pyproject.toml
# pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=77", "setuptools-scm>=8"]

[project]
name = "civic-geo-generator" # Package name
dynamic = ["version"]  # Using setuptools_scm
description = "Build, validate, and index web-optimized GeoJSON precinct, school district, etc. files for state repos"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name="Civic Interconnect", email="info@civic-interconnect.org" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Education",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: GIS",
  "Topic :: Software Development :: Build Tools",
  "Typing :: Typed",
]
keywords = [
  "civic", "elections", "gis", "geojson", "boundaries", "precincts", "generator", "publisher"
]
dependencies = [
  "civic-lib-core",
  "civic-lib-geo",
  "geopandas",
  "jsonschema[format]",
  "PyYAML",
  "typer",
]

[project.optional-dependencies]
dev = [ # Add all to deptry ignores
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest-env",
  "shapely",
  "twine",
  "validate-pyproject",
]
docs = [ # Add all to deptry ignores
  "mike",
  "mkdocs",
  "mkdocs-gen-files",
  "mkdocs-git-revision-date-localized-plugin",
  "mkdocs-include-markdown-plugin",
  "mkdocs-literate-nav",
  "mkdocs-material",
  "mkdocs-static-i18n",
  "mkdocstrings[python]>=0.18",
  "ruff", # required for mkdocstrings
]
# possible future split 
# geo = ["geopandas", "shapely", "pyogrio"]

[project.urls]
"Documentation" = "https://civic-interconnect.github.io/civic-geo-generator/latest"
"Homepage" = "https://github.com/civic-interconnect/civic-geo-generator"
"Issues" = "https://github.com/civic-interconnect/civic-geo-generator/issues"
"Source" = "https://github.com/civic-interconnect/civic-geo-generator"

[project.scripts]
civic-gen = "civic_geo_generator.cli.cli:main"

[project.entry-points."civic_geo_generator.adapters"]
# Adapters can register here; for now MN inside this package:
MN = "civic_geo_generator.adapters.mn:Adapter"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"] # B101: assert statements

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.paths]
source = [
  "src/",
  "*/site-packages/",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if settings.DEBUG",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod",
  "# pragma: no cover",
]
fail_under = 40
precision = 1
show_missing = true
skip_covered = false

[tool.coverage.run]
branch = true
omit = [
  "*/tests/*",
  "*/test_*",
  "*/__pycache__/*",
]
source = ["src"]

[tool.coverage.xml]
output = "coverage.xml"

[tool.deptry]
known_first_party = ["civic_geo_generator"]
package_module_name_map = {PyYAML = "yaml"  }

[tool.deptry.per_rule_ignores]
DEP002 = [
  # Dev dependencies
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest-env",
  "shapely",
  "twine",
  "validate-pyproject",
  # Docs dependencies
  "mike",
  "mkdocs",
  "mkdocs-gen-files",
  "mkdocs-git-revision-date-localized-plugin",
  "mkdocs-include-markdown-plugin",
  "mkdocs-literate-nav",
  "mkdocs-material",
  "mkdocs-static-i18n",
  "mkdocstrings",
  "ruff",
]

[tool.pyright]
executionEnvironments = [
  { root = ".",    extraPaths = ["./src"] },
  { root = "tests", extraPaths = ["./src"] },
]
extraPaths = ["./src"]
ignore = ["**/__pycache__", "build", "dist", ".venv"]
include = ["src", "tests"]
pythonVersion = "3.12"
reportMissingTypeStubs = "warning"
typeCheckingMode = "basic" # or "strict"
# reportReturnType = "none"

[tool.pytest.env]
PYTEST_CURRENT_TEST = "true"

[tool.pytest.ini_options]
addopts = [
  "-ra", # show extra test summary info for skipped, failed, etc.
  "--strict-markers", # enforce marker usage such as "unit", "integration", etc.
  "--cov=src", # measure coverage for files in src
  "--cov-report=term-missing:skip-covered", # term-missing:skip-covered is typical
  "--cov-report=xml:coverage.xml", # output coverage in XML format
  "--cov-report=html:htmlcov", # output coverage in HTML format
  "-v", # verbose output
  "--tb=short", # short traceback format
]
minversion = "7.0"
python_files = "test_*.py"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
exclude = [".venv", ".ruff_cache", "build", "dist", "*.egg-info", ".eggs", "__pycache__"]
fix = true
line-length = 100
preview = false
target-version = "py312"
unsafe-fixes = false

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"
quote-style = "preserve"

[tool.ruff.lint]
extend-select = [
  "F", # Pyflakes rules
  "W", # PyCodeStyle warnings
  "E", # PyCodeStyle errors
  "I", # Sort imports properly
  "UP", # Newer Python features
  "C4", # Catch incorrect use of comprehensions, dict, list, etc
  "C408", # unnecessary comprehension
  "D", # Docstring conventions
  "D401", # Missing argument descriptions in the docstring
  "D417", # First line should be in imperative mood (good habit)
  "Q", # flake8-quotes
  "S", # Security issues
  "C90", # McCabe complexity
  "N", # Naming conventions
  "B", # Bugbear
  "A", # flake8-annotations
  "PTH", # Use pathlib instead of os.path
  "RET", # Good return practices
  "SIM", # flake8-simplify
  "TID", # flake8-tidy-imports for project conventions
  "TCH", # flake8-type-checking for type checking imports
]
ignore = [
  "E501", # line length handled by formatter
  "D203", # 1 blank line before class docstring (conflicts with D211)
  "D213", # Multi-line docstring summary should start at the second line
  "S101", # assert statements (handled by bandit)
]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
"src/**/__init__.py" = ["D104"]
"tests/**/*.py" = ["TID251", "TID252", "S101", "D"]
"src/**/_version.py" = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"civic_geo_generator" = ["py.typed"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
fallback_version = "0.0.1"
local_scheme = "no-local-version"
version_file = "src/civic_geo_generator/_version.py"
version_scheme = "guess-next-dev"
